+++ src/Benchmarks/SharpNeat.Benchmarks/NeuralNets/NeuralNetCyclicBenchmarks.cs
@@ -10,20 +10,20 @@ namespace SharpNeat.NeuralNets;
 
 public class NeuralNetCyclicBenchmarks
 {
-    static readonly NeuralNetCyclic<double> __nn;
+    static readonly NeuralNetCyclic<float> __nn;
 
     static NeuralNetCyclicBenchmarks()
     {
         // Load neural net model from file, and convert into a neural net instance.
         NetFileModel netFileModel = NetFile.Load("data/genomes/preycapture.net");
-        __nn = (NeuralNetCyclic<double>)NeuralNetConverter.ToNeuralNet<double>(netFileModel);
+        __nn = (NeuralNetCyclic<float>)NeuralNetConverter.ToNeuralNet<float>(netFileModel);
 
         // Set some non-zero random input values.
         var rng = RandomDefaults.CreateRandomSource();
         var inputs = __nn.Inputs.Span;
         for(int i = 0; i < inputs.Length; i++)
         {
-            inputs[i] = rng.NextDouble();
+            inputs[i] = rng.NextFloat();
         }
     }
